improve API
